masters-worker/requirements.txt
# Racing API Reference Data Fetcher - Production Requirements
# Python 3.9+ required

# Core dependencies
python-dotenv>=1.0.0        # Environment variable management
requests>=2.31.0            # HTTP requests for Racing API
supabase>=2.0.0             # Supabase Python client

# Database
psycopg2-binary>=2.9.9      # PostgreSQL adapter
postgrest>=0.16.0           # PostgREST client (Supabase dependency)

# Utilities
python-dateutil>=2.8.2      # Date parsing and manipulation
schedule>=1.2.0             # Job scheduling (for Render.com worker)

# Optional but recommended for production
# sentry-sdk>=1.40.0        # Error tracking and monitoring
# prometheus-client>=0.19.0 # Metrics collection
